+++ docs/Whats-New.md
@@ -309,6 +309,13 @@ HideLightFlashEffects=false      ; boolean
 
 ## Changelog
 
+### 0.4.0.1
+
+```{dropdown} Click to show
+Vanilla fixes:
+- Fix preplaced aircraft outside visible map being incorrectly flagged as crashing under certain conditions (by Starkku)
+```
+
 ### 0.4
 
 ```{dropdown} Click to show